# *: optimize watch stream multiplexing garbage collection