@gyuho 2018-06-04
This still needs to handle gRPC proxy watch coalescing. WIP.