@gyuho 2018-06-15
Will reopen after fixing gRPC proxy test failures.